[KYUUBI #1542] Move FetchIterator into kyuubi-common for multi-engine support
